chore: remove re-export stubs and relocate shared components
Remove backward-compatibility re-export stubs:
- src/hooks/useLayoutRouting.ts → use @/features/layout-library
- src/hooks/useLayoutSwitcher.ts → use @/features/layout-library
- src/components/Modals/LayoutManagerModal/index.ts → use @/features/layout-library

Relocate domain-agnostic components to src/shared/components/:
- SectionHeader (used by settings panels)
- SettingsRow (used by settings panels)
- SelectDropdown (used by inspector, settings)
- BulkIncrementControl (used by multi-bin inspector)

Update all consumer imports to use new paths.
Update test mock paths for relocated modules.

All 3326 tests pass.
