--- src/components/Modals/LayoutManagerModal/index.ts
@@ -1,2 +1,0 @@
-// Re-export from new location for backward compatibility
-export { LayoutManagerModal } from '@/features/layout-library/components/LayoutManagerModal';
