--- src/hooks/useLayoutRouting.ts
@@ -1,2 +1,0 @@
-// Re-export from new location for backward compatibility
-export { useLayoutRouting } from '@/features/layout-library/hooks/useLayoutRouting';
