--- src/hooks/useLayoutSwitcher.ts
@@ -1,2 +1,0 @@
-// Re-export from new location for backward compatibility
-export { useLayoutSwitcher } from '@/features/layout-library/hooks/useLayoutSwitcher';
