+++ src/shared/components/BulkIncrementControl/index.ts
@@ -1,0 +1,1 @@
+export { BulkIncrementControl } from './BulkIncrementControl';
