+++ src/shared/components/SectionHeader/SectionHeader.tsx
