+++ src/shared/components/SectionHeader/index.ts
@@ -1,0 +1,1 @@
+export { SectionHeader } from './SectionHeader';
