+++ src/shared/components/SelectDropdown/index.ts
@@ -1,0 +1,1 @@
+export { SelectDropdown } from './SelectDropdown';
