+++ src/shared/components/SettingsRow/SettingsRow.tsx
