+++ src/shared/components/SettingsRow/index.ts
@@ -1,0 +1,1 @@
+export { SettingsRow } from './SettingsRow';
