+++ src/test/hooks/useLayoutRouting-cloudShare.test.ts
@@ -8,7 +8,7 @@
 
 import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest';
 import { renderHook, act } from '@testing-library/react';
-import { useLayoutRouting } from '@/hooks/useLayoutRouting';
+import { useLayoutRouting } from '@/features/layout-library/hooks/useLayoutRouting';
 import { useLayoutStore } from '@/core/store/layout';
 import { useLibraryStore } from '@/core/store/library';
 import { useUIStore } from '@/core/store/ui';
