+++ src/test/hooks/useLayoutRouting.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest';
 import { renderHook, act } from '@testing-library/react';
-import { useLayoutRouting } from '@/hooks/useLayoutRouting';
+import { useLayoutRouting } from '@/features/layout-library/hooks/useLayoutRouting';
 import { useLayoutStore, useLibraryStore, useUIStore, useToastStore } from '@/core/store';
 import { resetAllStores } from '@/test/testUtils';
 import * as storage from '@/core/storage';
